# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Personal Finance API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtener un usuario específico con sus carteras",
                "operationId": "0197efcfd303b59ccdaad3e01097a0e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del usuario y sus carteras",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "wallet": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Wallet"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/api/v1/users/wallets": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Obtener todas las carteras del usuario autenticado",
                "operationId": "257be872852858814e0df7507dcebadb",
                "responses": {
                    "200": {
                        "description": "Lista de carteras del usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "wallets": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "wallet_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "account_type": {
                                                        "type": "string",
                                                        "example": "Bancaria"
                                                    },
                                                    "balance": {
                                                        "type": "number",
                                                        "example": 1000
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-06T19:10:44Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-06T19:10:44Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay carteras creadas en este usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No hay carteras creadas en este usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error inesperado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error inesperado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/wallets/create": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Crear una nueva cartera para el usuario autenticado",
                "operationId": "293c8c2a511ec37cda317696847aeb79",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type_account": {
                                        "type": "string",
                                        "example": "Bancaria"
                                    },
                                    "balance": {
                                        "type": "number",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cartera creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "wallet creada correctamente": {
                                            "properties": {
                                                "wallet_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "account_type": {
                                                    "type": "string",
                                                    "example": "Bancaria"
                                                },
                                                "balance": {
                                                    "type": "number",
                                                    "example": 5000
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-06T19:10:44Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-06T19:10:44Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error inesperado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error inesperado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Wallet": {
                "description": "Wallet model",
                "properties": {
                    "id": {
                        "description": "Wallet ID",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "User ID associated with the wallet",
                        "type": "integer"
                    },
                    "account_type_id": {
                        "description": "Account type ID associated with the wallet",
                        "type": "integer"
                    },
                    "balance": {
                        "description": "Wallet balance",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Se requiere un token de acceso Bearer JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}